sched: Allow scheduler to use UCLAMP min percentage for schedtune SPC boosting
This will allow us to configure the scheduler to use the UCLAMP min
value percentage over SCHED_CAPACITY_SCALE for schedtune SPC boosting.

SPC boosting greatly benefits older devices wherein there is a need
to compensate for important (top-app) tasks as we want the CPU to run
at a high frequency while executing these tasks, or to place them on a
higher capacity CPU, to ensure lower latencies as its lower frequencies
or efficient CPUs are not sufficient. In this case, we will be applying
SPC boosting for tasks with util above the UCLAMP_MIN to boost tasks
which need the compensation relative to how far they are from the CPU
capacity.

This feature requires both SCHED_TUNE and UCLAMP_TASK to be supported
and enabled.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>